Keypad / DRO improvement
I added a basic function "dashboard.DRO()" that can be called from the
dashboard or an app.
For the moment, this function :

Display the right menu (that itself make the keypad usable)
Give an instruction to the user
Wait keypress "enter" to call the callback.
I guess there is others things to do on this function.